requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
twilio==8.10.0
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.0
python-multipart==0.0.6




